requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
Pillow>=9.0.0
psutil>=5.9.0
streamlit>=1.28.0
